Projeto finalizado

+++ packages/backend/services/stock/internal/models/product.go
@@ -12,14 +12,14 @@ type Product struct {
 	ID        uuid.UUID      `json:"id" gorm:"type:uuid;default:gen_random_uuid();primaryKey"`
 	Name      string         `json:"name" binding:"required"`
 	SKU       string         `json:"sku" binding:"required"`
-	Quantity  int            `json:"quantity" binding:"gte=0"`
-	Price     float64        `json:"price" binding:"gt=0"`
+	Quantity  int64          `json:"quantity" binding:"gte=0"`
+	Price     int64          `json:"price" binding:"gt=0"`
 	CreatedAt time.Time      `json:"created_at"`
 	UpdatedAt time.Time      `json:"updated_at"`
 	DeletedAt gorm.DeletedAt `json:"-" gorm:"index" swaggerignore:"true"`
 }
 
-func (p *Product) ReduceStock(qty int) error {
+func (p *Product) ReduceStock(qty int64) error {
 	if qty <= 0 {
 		return apperrors.ErrInvalidInput
 	}
@@ -30,7 +30,7 @@ func (p *Product) ReduceStock(qty int) error {
 	return nil
 }
 
-func (p *Product) RestoreStock(qty int) error {
+func (p *Product) RestoreStock(qty int64) error {
 	if qty <= 0 {
 		return apperrors.ErrInvalidInput
 	}
